+++ app/src/components/ListLayer.vue
@@ -36,18 +36,23 @@
           </div>
 
           <div class="btn-row">
-            <BaseButton buttonClass="danger" @click="deleteFlag()"
+            <!-- <BaseButton buttonClass="danger" @click="deleteFlag()"
               >Discard</BaseButton
-            >
+            > -->
+            <SvgButton buttonClass="nodes" @click.prevent="deleteFlag()" />
             <div v-if="value.read_mode == true && deleted == false">
-              <BaseButton class="read" buttonClass="action" @click="readFlag()"
+              <SvgButton2 buttonClass="nodes" @click.prevent="readFlag()" />
+
+              <!-- <BaseButton class="read" buttonClass="action" @click="readFlag()"
                 >Edit Mode
-              </BaseButton>
+              </BaseButton> -->
             </div>
             <div v-else>
-              <BaseButton class="read" buttonClass="action" @click="readFlag()"
+              <SvgButton2 buttonClass="nodes" @click.prevent="readFlag()" />
+
+              <!-- <BaseButton class="read" buttonClass="action" @click="readFlag()"
                 >Read Mode</BaseButton
-              >
+              > -->
             </div>
           </div>
         </form>
@@ -59,6 +64,9 @@
 <script>
 import { mapState } from 'vuex'
 import marked from 'marked'
+import SvgButton from '@/components/SvgButton'
+import SvgButton2 from '@/components/SvgButton2'
+
 var readmode
 export default {
   name: 'ListLayer',
@@ -131,6 +139,10 @@ export default {
         this.mode = 'Edit'
       }
     },
+  },
+  components: {
+    SvgButton,
+    SvgButton2,
   },
 }
 </script>